net: wireless: rockchip_wlan: update rtl8188fu wifi driver
wifi driver version: v4.3.23.4_19019.20160811

+++ drivers/net/wireless/rockchip_wlan/rtl8188fu/include/sta_info.h
@@ -242,7 +242,7 @@ struct sta_info {
 	u8	SNonce[32];
 	u8	ANonce[32];
 	u32	TDLS_PeerKey_Lifetime;
-	u16	TPK_count;
+	u32	TPK_count;
 	_timer	TPK_timer;
 	struct TDLS_PeerKey	tpk;
 #ifdef CONFIG_TDLS_CH_SW
